chore: rename create-realm-validator
